Fixed #21 : Properly handle long values that are greater than an int value.

+++ core/src/main/java/com/dexesttp/hkxpack/resources/byteutils/ULongByteUtils.java
@@ -22,7 +22,7 @@ final class ULongByteUtils {
 		long accu = 1;
 		long res = 0;
 		for (int i = 0; i < len; i++) {
-			res += ((int) (list[i] & 0xFF)) * accu;
+			res += ((long) (list[i] & 0xFF)) * accu;
 			accu *= 256;
 		}
 		return res;
